+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Add "additionalInformation" and "officeHoursEmployee" attributes to PersonClaimsResource
+
 ## v0.3.34
 
 - public rest resource objects: Standardize methods names for localized attributes (e.g. getName() and getNameLocalized())